Create elr.json
Based on work parsing sample ELR messages. A few considerations:

* There was no `Person` record, which was a required field for ECR. I did not see an equivalent person ID in these fields:

```python
"""I put the full json into python to get a count of reach resourceType; there are numerous repeating ones to try to make sense of
Counter({'Location': 68,
         'Organization': 26,
         'Practitioner': 23,
         'PractitionerRole': 14,
         'Device': 6,
         'DiagnosticReport': 3,
         'Specimen': 3,
         'ServiceRequest': 2,
         'Observation': 2,
         'MessageHeader': 1,
         'Provenance': 1,
         'Patient': 1,
         'EpisodeOfCare': 1,
         'Encounter': 1,
         'DocumentReference': 1,
         'ResearchStudy': 1,
         'ChargeItem': 1,
         'Procedure': 1,
         'MedicationDispense': 1})
 """
```

* Do we need to dedupe duplicate Location/Organizations in some way? These are the largest, broadly duplicated records. Another thought I had, based at least on the sample I was working with, was to try to link "backwards"; for example, a lot of recordtypes had a link to organization, practitioner records, etc.

I am wondering if for ELR data is makes sense to essentially work backwards: start with a specimen and then link back to the relevant observation, location recordtype with the relevant info there. I was not entirely sure the best way to do that within the json, or if that would make sense for all ELR data. I'm happy to keep experimenting with sample ELR / see the best way to do that kind of dynamic link-back.