remove redundant initialize schema calls in dal.py
The dal::get_table_by_column_* methods do not need to invoke the
initialize_schema call every time they are used, it should be sufficient
to just call that method with the DAL instance is created. The MPI
connector client uses these methods when building a block query,
initializing the SQLAlchemy schema every time a block query needs to
be generated adds overhead.