replace aggregate given names function with query
Using pandas dataframes has some overhead to it, in some cases adding up
to an extra 100ms.  Since we are already using the database query to
retrieve the data, we can amend the original query to do the aggregation
for us.  The updated query in mpi.py, aggregates the given_names into a
list of values, so we still need to convert that into a white space
delimited value in link.py, but pandas is not needed for that.