record linkage performance enhancments (#1651)
* remove redundant initialize schema calls in dal.py

The dal::get_table_by_column_* methods do not need to invoke the
initialize_schema call every time they are used, it should be sufficient
to just call that method with the DAL instance is created. The MPI
connector client uses these methods when building a block query,
initializing the SQLAlchemy schema every time a block query needs to
be generated adds overhead.

* optimize FK name lookup when creating block query

The FK table and column name can be retrieved from the CTE foreign keys
set by calling `next(iter(...))`, rather than making a copy and popping
the first item off.  The latter has an added expense of re initializing a
SQLAlchemy table object, which can on occasion add up to 100ms.

* add link::_convert_given_name_to_first_name

For transforming a dataset with a given name column (that contains a
list of names) to a first name column (that contains a white space
delimited name)

* replace aggregate given names function with query

Using pandas dataframes has some overhead to it, in some cases adding up
to an extra 100ms.  Since we are already using the database query to
retrieve the data, we can amend the original query to do the aggregation
for us.  The updated query in mpi.py, aggregates the given_names into a
list of values, so we still need to convert that into a white space
delimited value in link.py, but pandas is not needed for that.

* [pre-commit.ci] auto fixes from pre-commit hooks

* update test_mpi cases to meet the requirements of the new base query

* update test_linkage cases in containers

* [pre-commit.ci] auto fixes from pre-commit hooks

* run distinct on block query CTE

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>